// src/API/AutoShopApi/models/CarDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AutoShopAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UserBasicDTO } from './UserBasicDTO';
import {
    UserBasicDTOFromJSON,
    UserBasicDTOFromJSONTyped,
    UserBasicDTOToJSON,
    UserBasicDTOToJSONTyped,
} from './UserBasicDTO';

/**
 * 
 * @export
 * @interface CarDTO
 */
export interface CarDTO {
    /**
     * 
     * @type {number}
     * @memberof CarDTO
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof CarDTO
     */
    company?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CarDTO
     */
    model?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof CarDTO
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof CarDTO
     */
    updatedAt?: Date;
    /**
     * 
     * @type {Array<UserBasicDTO>}
     * @memberof CarDTO
     */
    users?: Array<UserBasicDTO> | null;
}

/**
 * Check if a given object implements the CarDTO interface.
 */
export function instanceOfCarDTO(value: object): value is CarDTO {
    return true;
}

export function CarDTOFromJSON(json: any): CarDTO {
    return CarDTOFromJSONTyped(json, false);
}

export function CarDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): CarDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'company': json['company'] == null ? undefined : json['company'],
        'model': json['model'] == null ? undefined : json['model'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'updatedAt': json['updatedAt'] == null ? undefined : (new Date(json['updatedAt'])),
        'users': json['users'] == null ? undefined : ((json['users'] as Array<any>).map(UserBasicDTOFromJSON)),
    };
}

export function CarDTOToJSON(json: any): CarDTO {
    return CarDTOToJSONTyped(json, false);
}

export function CarDTOToJSONTyped(value?: CarDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'company': value['company'],
        'model': value['model'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
        'updatedAt': value['updatedAt'] == null ? undefined : ((value['updatedAt']).toISOString()),
        'users': value['users'] == null ? undefined : ((value['users'] as Array<any>).map(UserBasicDTOToJSON)),
    };
}