// src/API/AutoShopApi/models/UserBasicDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AutoShopAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UserBasicDTO
 */
export interface UserBasicDTO {
    /**
     * 
     * @type {number}
     * @memberof UserBasicDTO
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof UserBasicDTO
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserBasicDTO
     */
    email?: string | null;
}

/**
 * Check if a given object implements the UserBasicDTO interface.
 */
export function instanceOfUserBasicDTO(value: object): value is UserBasicDTO {
    return true;
}

export function UserBasicDTOFromJSON(json: any): UserBasicDTO {
    return UserBasicDTOFromJSONTyped(json, false);
}

export function UserBasicDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserBasicDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'email': json['email'] == null ? undefined : json['email'],
    };
}

export function UserBasicDTOToJSON(json: any): UserBasicDTO {
    return UserBasicDTOToJSONTyped(json, false);
}

export function UserBasicDTOToJSONTyped(value?: UserBasicDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'email': value['email'],
    };
}